+++ src/lib/pages/account-details/components/AccountTitle.tsx
@@ -9,19 +9,21 @@ import type { Option } from "lib/types";
 interface AccountTitleProps {
   accountData: Option<AccountData>;
   accountLocalInfo: Option<AccountLocalInfo>;
-  data?: {
-    username: string | null;
-  };
-  isLoading?: boolean;
-  isFetching?: boolean;
+  initiaUsernameData: Option<InitiaUsernameDataResponse>;
+  isInitiaUsernameDataLoading: boolean;
+  isInitiaUsernameDataFetching: boolean;
+}
+
+export interface InitiaUsernameDataResponse {
+  username: string | null;
 }
 
 export const AccountTitle = ({
   accountData,
   accountLocalInfo,
-  data,
-  isLoading,
-  isFetching,
+  initiaUsernameData,
+  isInitiaUsernameDataLoading,
+  isInitiaUsernameDataFetching,
 }: AccountTitleProps) => {
   const move = useMoveConfig({ shouldRedirect: false });
 
@@ -29,7 +31,8 @@ export const AccountTitle = ({
     if (accountLocalInfo?.name) return accountLocalInfo.name;
     if (accountData?.publicInfo?.name) return accountData?.publicInfo?.name;
     if (accountData?.icns?.primaryName) return accountData?.icns?.primaryName;
-    if (move.enabled && data?.username) return data?.username;
+    if (move.enabled && initiaUsernameData?.username)
+      return initiaUsernameData?.username;
     return "Account Details";
   };
 
@@ -48,7 +51,7 @@ export const AccountTitle = ({
         />
       );
 
-    if (move.enabled && data?.username && !accountLocalInfo?.name)
+    if (move.enabled && initiaUsernameData?.username && !accountLocalInfo?.name)
       return (
         <Image
           src="https://assets.alleslabs.dev/webapp-assets/name-services/initia-username.svg"
@@ -65,7 +68,7 @@ export const AccountTitle = ({
     return <CustomIcon name="wallet" boxSize={5} color="secondary.main" />;
   };
 
-  if (isLoading && isFetching)
+  if (isInitiaUsernameDataLoading && isInitiaUsernameDataFetching)
     return (
       <Skeleton
         h={6}