+++ src/lib/services/tx/index.ts
@@ -20,6 +20,7 @@ import {
   useValidateAddress,
   useWasmConfig,
 } from "lib/app-provider";
+import { createQueryFnWithTimeout } from "lib/query-utils";
 import type {
   BechAddr,
   BechAddr20,
@@ -281,30 +282,6 @@ export const useTxsByContractAddressLcd = (
   );
 };
 
-export const useTxsByAccountAddressLcd = (
-  address: Option<BechAddr20>,
-  limit: number,
-  offset: number
-) => {
-  const endpoint = useLcdEndpoint();
-
-  return useQuery(
-    [
-      CELATONE_QUERY_KEYS.TXS_BY_ACCOUNT_ADDRESS_LCD,
-      endpoint,
-      address,
-      limit,
-      offset,
-    ],
-    async () => {
-      if (!address)
-        throw new Error("address is undefined (useTxsByAccountAddressLcd)");
-      return getTxsByAccountAddressLcd(endpoint, address, limit, offset);
-    },
-    { retry: 1, refetchOnWindowFocus: false }
-  );
-};
-
 export const useTxsByAddressLcd = (
   address: Option<BechAddr20>,
   search: Option<string>,
@@ -361,7 +338,7 @@ export const useTxsByAddressLcd = (
       limit,
       offset,
     ],
-    queryfn,
+    createQueryFnWithTimeout(queryfn, 20000),
     { ...options, retry: 1, refetchOnWindowFocus: false }
   );
 };